+++ allennlp/data/vocabulary.py
@@ -9,6 +9,7 @@ import logging
 import os
 import re
 from collections import defaultdict
+from transformers import PreTrainedTokenizer
 from typing import Any, Callable, Dict, Iterable, List, Optional, Set, Union, TYPE_CHECKING
 
 from allennlp.common import Registrable
@@ -16,7 +17,7 @@ from allennlp.common.file_utils import cached_path, FileLock
 from allennlp.common.checks import ConfigurationError
 from allennlp.common.tqdm import Tqdm
 from allennlp.common.util import namespace_match
-from allennlp.tools import copy_transformer_vocab
+
 
 if TYPE_CHECKING:
     from allennlp.data import instance as adi  # noqa
@@ -29,6 +30,26 @@ DEFAULT_PADDING_TOKEN = "@@PADDING@@"
 DEFAULT_OOV_TOKEN = "@@UNKNOWN@@"
 NAMESPACE_PADDING_FILE = "non_padded_namespaces.txt"
 _NEW_LINE_REGEX = re.compile(r"\n|\r\n")
+
+
+def copy_transformer_vocab(tokenizer: PreTrainedTokenizer) -> Dict[str, Any]:
+    """
+    Copies tokens from ```transformers``` model's vocab
+    """
+    try:
+        vocab_items = tokenizer.get_vocab().items()
+    except NotImplementedError:
+        vocab_items = (
+            (tokenizer.convert_ids_to_tokens(idx), idx) for idx in range(tokenizer.vocab_size)
+        )
+    outputs = dict()
+    outputs["token_to_index"] = dict()
+    outputs["index_to_token"] = dict()
+    for word, idx in vocab_items:
+        outputs["token_to_index"][word] = idx
+        outputs["index_to_token"][idx] = word
+
+    return outputs
 
 
 class _NamespaceDependentDefaultDict(defaultdict):
